+++ src/components/spartan/SRadioGroup/types.ts
@@ -1,0 +1,10 @@
+export type TRadioGroupEmits = {
+    (event: 'update:modelValue', value: string): void;
+};
+export type TRadioGroupProps = {
+    modelValue: string;
+};
+
+export type TRadioGroupItemProps = {
+    value: string;
+};